Move media type code to another module.
